Move CPE handling methods out of Player.Networking.cs
